make duplicate + rename actions work on non-local layer stack layers and change behavior to target the edit target layer for the new prims.
Some more thought needs to be put into targeting layers, but at least
this gets the actions working for stages with references. Editing usd
stages that are only single layers is pretty limited, but these changes
will be robust for that case if you set the edit target to the root layer.

Duplicate is not as big of a problem and could be made even more robust
by flattening prim specs in the copy and expecting the edit target (session
layer by default) to be flattened onto root layer at export time.

Rename is trickier and ive left some notes in the code.